move writeHead array tests to a dedicated file; clear raw snapshot before ERR_HTTP_TRAILER_INVALID
node-http.test.ts has a pre-existing container-specific failure
(request via http proxy, issue#4295: localhost resolves to ::1) that
blocks the gate's clean-pass check; a dedicated file isolates the
fail-before/pass-after proof.

Also clear kRawOutHeaders before throwing ERR_HTTP_TRAILER_INVALID so a
handler that catches the error and retries does not render the rejected
snapshot.
